# api/openapi/hakoniwa.yaml
openapi: 3.0.3
info:
  title: Hakoniwa API
  version: 0.1.0
  description: API for Hakoniwa On-Demand Desktop Service
servers:
  - url: /_hakoniwa/api
paths:
  /auth/me:
    get:
      summary: Get current user and instance status
      operationId: getAuthMe
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthStatus'
        '401':
          description: Unauthorized (Not logged in)
  /auth/anonymous:
    post:
      summary: Login anonymously (creates session and instance)
      operationId: loginAnonymous
      responses:
        '200':
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthStatus'
  /auth/oidc/authorize:
    get:
      summary: Redirect to OIDC provider
      operationId: oidcAuthorize
      responses:
        '302':
          description: Redirect to OIDC provider
          headers:
            Location:
              schema:
                type: string
              description: URL to redirect to
              required: true
  /auth/oidc/callback:
    get:
      summary: Process OIDC callback from IdP
      operationId: oidcCallback
      parameters:
        - name: code
          in: query
          schema:
            type: string
          required: true
        - name: state
          in: query
          schema:
            type: string
          required: true
        - name: error
          in: query
          schema:
            type: string
          required: false
          description: Error code from IdP
        - name: error_description
          in: query
          schema:
            type: string
          required: false
          description: Error description from IdP
      responses:
        '302':
          description: Redirect to frontend
          headers:
            Location:
              schema:
                type: string
              description: URL to redirect to
              required: true
  /configuration:
    get:
      summary: Get application configuration
      operationId: getConfiguration
      responses:
        '200':
          description: Application configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Configuration'

components:
  schemas:
    Configuration:
      type: object
      properties:
        title:
          type: string
          description: Application title
        message:
          type: string
          description: Welcome message
        logo_url:
          type: string
          description: URL to the application logo
        terms_of_service_url:
          type: string
          description: URL to the terms of service
        privacy_policy_url:
          type: string
          description: URL to the privacy policy
        auth_methods:
          type: array
          items:
            type: string
          description: List of enabled authentication methods (e.g. 'anonymous', 'oidc')
        oidc_name:
          type: string
          description: Display name for the OIDC login button
          default: OpenID Connect
        auth_auto_login:
          type: boolean
          description: Whether to automatically log in if only one auth method is available
          default: false
      required:
        - title
        - message
        - logo_url
        - auth_methods
        - auth_auto_login
    AuthStatus:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        instance:
          $ref: '#/components/schemas/Instance'
      required:
        - user
    User:
      type: object
      properties:
        id:
          type: string
          description: User ID (OpenID Connect sub or UUID)
        type:
          type: string
          enum: [openid_connect, anonymous]
      required:
        - id
        - type
    Instance:
      type: object
      properties:
        status:
          type: string
          enum: [pending, running, terminating]
        pod_ip:
          type: string
      required:
        - status